src/test/java/dev/dobicinaitis/advent/Day5Test.java
package dev.dobicinaitis.advent;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Day5Test {

    private Puzzle puzzle = new Day5();

    @Test
    public void firstPuzzleExampleTest() {
        assertEquals("CMZ", puzzle.getFirstPuzzleSolution(puzzle.exampleInputFilename));
    }

    @Test
    public void firstPuzzleMyInputTest() {
        assertEquals("BWNCQRMDB", puzzle.getFirstPuzzleSolution(puzzle.myInputFilename));
    }

    @Test
    public void secondPuzzleExampleTest() {
        assertEquals("MCD", puzzle.getSecondPuzzleSolution(puzzle.exampleInputFilename));
    }

    @Test
    public void secondPuzzleMyInputTest() {
        assertEquals("NHWZCBNBF", puzzle.getSecondPuzzleSolution(puzzle.myInputFilename));
    }
}